Move type definitions out of config.ts
